+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Unit tests for backend.py file
+
 ## [0.13.1] - 2020-03-01
 ### Added
 - example code using the API (sample.py)
@@ -168,6 +172,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic server and client
 - Use regex to parse commands
 
+[Unreleased]: https://github.com/RomainL972/isn/compare/v0.13.1...HEAD
 [0.13.1]: https://github.com/RomainL972/isn/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/RomainL972/isn/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/RomainL972/isn/compare/v0.11.0...v0.12.0